+++ btc_cycles/halvings.json
@@ -3,7 +3,7 @@
   "210000": { "reward": 25.0, "date": "2012-11-28T00:00:00Z" },
   "420000": { "reward": 12.5, "date": "2016-07-09T00:00:00Z" },
   "630000": { "reward": 6.25, "date": "2020-05-11T00:00:00Z" },
-  "840000": { "reward": 3.125, "date": "2024-04-24T12:09:27Z" },
+  "840000": { "reward": 3.125, "date": "2024-04-20T12:09:27Z" },
   "1050000": { "reward": 1.5625, "date": "" },
   "1260000": { "reward": 0.78125, "date": "" },
   "1470000": { "reward": 0.390625, "date": "" },
